Implemented a detailed service information display screen featuring a structured table layout for service details and modified the layout for enhanced readability and user interaction.
