+++ lib/main.dart
@@ -1,4 +1,4 @@
-import 'package:car_pulse/screens/service_book.dart';
+import 'package:car_pulse/screens/addServiceInfo.dart';
 import 'package:flutter/material.dart';
 
 void main() {
